pyproject.toml
[project]
name = "bankfetch"
version = "0.1.0"
description = "Download your own bank accounts transactions via the Enable Banking API"
readme = "README.md"
requires-python = ">=3.13"
license = { file = "LICENSE" }
authors = [{ name = "Davide Mungo", email = "dp.mungo@gmail.com" }]
keywords = [
    "banking", "transactions", "enable-banking", "open-banking",
    "bank-transactions", "personal-finance", "ledger",
    "download", "export", "csv", "cli", "psd2", "aisp"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Office/Business :: Financial",
    "License :: Other/Proprietary License",
    "Intended Audience :: End Users/Desktop",
]
dependencies = [
    "requests>=2.32",
    "python-dotenv>=1.0",
    "PyJWT>=2.8",
    "cryptography>=42.0",
    "typer>=0.12",
    "loguru>=0.7",
]

[project.urls]
Repository = "https://github.com/dpmungo/bankfetch"

[project.scripts]
bankfetch = "bankfetch.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/bankfetch"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.0",
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "ty>=0.0.0a8",
]
